+++ lib/presentation/router/app_router.dart
@@ -1,4 +1,5 @@
 import 'package:flutter/material.dart';
+import 'package:flutter_practice/presentation/screens/my_form_screen.dart';
 import 'package:flutter_practice/presentation/screens/random_jokes_screen.dart';
 import 'package:flutter_practice/presentation/screens/calculator_screen.dart';
 import 'package:flutter_practice/presentation/screens/home_screen.dart';
@@ -8,7 +9,7 @@ import '../../core/exceptions/route_exception.dart';
 
 class AppRouter {
   static const String splash = '/',
-      home = 'home', calc = 'calc', randomJokes = 'randomJokes';
+      home = 'home', calc = 'calc', randomJokes = 'randomJokes', myForm = 'myForm';
 
   const AppRouter._();
 
@@ -29,6 +30,10 @@ class AppRouter {
       case randomJokes:
         return MaterialPageRoute(
           builder: (_) => const RandomJokesScreen(),
+        );
+      case myForm:
+        return MaterialPageRoute(
+          builder: (_) => const MyFormScreen(),
         );
       default:
         throw const RouteException('Route not found!');